ai-learn/pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.1",
    "openai>=1.88.0",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
    "requests>=2.32.4",
]
